+++ extensions/schema-compare/src/dialogs/schemaCompareDialog.ts
@@ -11,19 +11,23 @@ import * as os from 'os';
 import { SchemaCompareResult } from '../schemaCompareResult';
 
 const localize = nls.loadMessageBundle();
-const CompareButtonText: string = localize('schemaCompareDialog.Compare', 'Compare');
-const CancelButtonText: string = localize('schemaCompareDialog.Cancel', 'Cancel');
-const SourceTextBoxLabel: string = localize('schemaCompareDialog.SourceLabel', 'Source File');
-const TargetTextBoxLabel: string = localize('schemaCompareDialog.TargetLabel', 'Target File');
+const OkButtonText: string = localize('schemaCompareDialog.ok', 'Ok');
+const CancelButtonText: string = localize('schemaCompareDialog.cancel', 'Cancel');
+const SourceTitle: string = localize('schemaCompareDialog.SourceTitle', 'Source');
+const TargetTitle: string = localize('schemaCompareDialog.TargetTitle', 'Target');
+const FileTextBoxLabel: string = localize('schemaCompareDialog.fileTextBoxLabel', 'File');
 const DacpacRadioButtonLabel: string = localize('schemaCompare.dacpacRadioButtonLabel', 'Data-tier Application File (.dacpac)');
 const DatabaseRadioButtonLabel: string = localize('schemaCompare.databaseButtonLabel', 'Database');
-const SourceRadioButtonsLabel: string = localize('schemaCompare.sourceButtonsLabel', 'Source Type');
-const TargetRadioButtonsLabel: string = localize('schemaCompare.targetButtonsLabel', 'Target Type');
-const NoActiveConnectionsLabel: string = localize('schemaCompare.NoActiveConnectionsText', 'No active connections');
+const RadioButtonsLabel: string = localize('schemaCompare.radioButtonsLabel', 'Type');
+const ServerDropdownLabel: string = localize('schemaCompareDialog.serverDropdownTitle', 'Server');
+const DatabaseDropdownLabel: string = localize('schemaCompareDialog.databaseDropdownTitle', 'Database');
+const NoActiveConnectionsLabel: string = localize('schemaCompare.noActiveConnectionsText', 'No active connections');
 const SchemaCompareLabel: string = localize('schemaCompare.dialogTitle', 'Schema Compare');
+const titleFontSize: number = 13;
 
 export class SchemaCompareDialog {
 	public dialog: azdata.window.Dialog;
+	public dialogName: string;
 	private schemaCompareTab: azdata.window.DialogTab;
 	private sourceDacpacComponent: azdata.FormComponent;
 	private sourceTextBox: azdata.InputBoxComponent;
@@ -45,7 +49,7 @@ export class SchemaCompareDialog {
 	private sourceIsDacpac: boolean;
 	private targetIsDacpac: boolean;
 	private database: string;
-	public dialogName: string;
+	private connectionId: string;
 
 	protected initializeDialog(): void {
 		this.schemaCompareTab = azdata.window.createTab(SchemaCompareLabel);
@@ -53,10 +57,17 @@ export class SchemaCompareDialog {
 		this.dialog.content = [this.schemaCompareTab];
 	}
 
-	public openDialog(p: any, dialogName?: string): void {
-		let profile = p ? <azdata.IConnectionProfile>p.connectionProfile : undefined;
+	public async openDialog(context: any, dialogName?: string): Promise<void> {
+		let profile = context ? <azdata.IConnectionProfile>context.connectionProfile : undefined;
 		if (profile) {
 			this.database = profile.databaseName;
+			this.connectionId = profile.id;
+		} else {
+			let connection = await azdata.connection.getCurrentConnection();
+			if (connection) {
+				this.connectionId = connection.connectionId;
+				this.database = undefined;
+			}
 		}
 
 		let event = dialogName ? dialogName : null;
@@ -64,7 +75,7 @@ export class SchemaCompareDialog {
 
 		this.initializeDialog();
 
-		this.dialog.okButton.label = CompareButtonText;
+		this.dialog.okButton.label = OkButtonText;
 		this.dialog.okButton.onClick(async () => await this.execute());
 
 		this.dialog.cancelButton.label = CancelButtonText;
@@ -81,7 +92,7 @@ export class SchemaCompareDialog {
 		if (this.sourceIsDacpac) {
 			sourceName = this.sourceTextBox.value;
 			sourceEndpointInfo = {
-				endpointType: azdata.SchemaCompareEndpointType.dacpac,
+				endpointType: azdata.SchemaCompareEndpointType.Dacpac,
 				serverName: '',
 				databaseName: '',
 				ownerUri: '',
@@ -92,7 +103,7 @@ export class SchemaCompareDialog {
 			let ownerUri = await azdata.connection.getUriForConnection((this.sourceServerDropdown.value as ConnectionDropdownValue).connection.connectionId);
 
 			sourceEndpointInfo = {
-				endpointType: azdata.SchemaCompareEndpointType.database,
+				endpointType: azdata.SchemaCompareEndpointType.Database,
 				serverName: (this.sourceServerDropdown.value as ConnectionDropdownValue).name,
 				databaseName: (<azdata.CategoryValue>this.sourceDatabaseDropdown.value).name,
 				ownerUri: ownerUri,
@@ -104,7 +115,7 @@ export class SchemaCompareDialog {
 		if (this.targetIsDacpac) {
 			targetName = this.targetTextBox.value;
 			targetEndpointInfo = {
-				endpointType: azdata.SchemaCompareEndpointType.dacpac,
+				endpointType: azdata.SchemaCompareEndpointType.Dacpac,
 				serverName: '',
 				databaseName: '',
 				ownerUri: '',
@@ -115,7 +126,7 @@ export class SchemaCompareDialog {
 			let ownerUri = await azdata.connection.getUriForConnection((this.targetServerDropdown.value as ConnectionDropdownValue).connection.connectionId);
 
 			targetEndpointInfo = {
-				endpointType: azdata.SchemaCompareEndpointType.database,
+				endpointType: azdata.SchemaCompareEndpointType.Database,
 				serverName: (this.targetServerDropdown.value as ConnectionDropdownValue).name,
 				databaseName: (<azdata.CategoryValue>this.targetDatabaseDropdown.value).name,
 				ownerUri: ownerUri,
@@ -169,23 +180,43 @@ export class SchemaCompareDialog {
 			if (this.database) {
 				this.formBuilder = view.modelBuilder.formContainer()
 					.withFormItems([
-						sourceRadioButtons,
-						this.sourceServerComponent,
-						this.sourceDatabaseComponent,
-						targetRadioButtons,
-						this.targetDacpacComponent
+						{
+							title: SourceTitle,
+							components: [
+								sourceRadioButtons,
+								this.sourceServerComponent,
+								this.sourceDatabaseComponent
+							]
+						}, {
+							title: TargetTitle,
+							components: [
+								targetRadioButtons,
+								this.targetDacpacComponent
+							]
+						}
 					], {
-							horizontal: true
+							horizontal: true,
+							titleFontSize: titleFontSize
 						});
 			} else {
 				this.formBuilder = view.modelBuilder.formContainer()
 					.withFormItems([
-						sourceRadioButtons,
-						this.sourceDacpacComponent,
-						targetRadioButtons,
-						this.targetDacpacComponent
+						{
+							title: SourceTitle,
+							components: [
+								sourceRadioButtons,
+								this.sourceDacpacComponent,
+							]
+						}, {
+							title: TargetTitle,
+							components: [
+								targetRadioButtons,
+								this.targetDacpacComponent
+							]
+						}
 					], {
-							horizontal: true
+							horizontal: true,
+							titleFontSize: titleFontSize
 						});
 			}
 			let formModel = this.formBuilder.component();
@@ -232,7 +263,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: currentTextbox,
-			title: isTarget ? TargetTextBoxLabel : SourceTextBoxLabel,
+			title: FileTextBoxLabel,
 			actions: [currentButton]
 		};
 	}
@@ -256,17 +287,17 @@ export class SchemaCompareDialog {
 			this.formBuilder.removeFormItem(this.sourceNoActiveConnectionsText);
 			this.formBuilder.removeFormItem(this.sourceServerComponent);
 			this.formBuilder.removeFormItem(this.sourceDatabaseComponent);
-			this.formBuilder.insertFormItem(this.sourceDacpacComponent, 1, { horizontal: true });
+			this.formBuilder.insertFormItem(this.sourceDacpacComponent, 2, { horizontal: true, titleFontSize: titleFontSize });
 		});
 
 		// show server and db dropdowns or 'No active connections' text
 		databaseRadioButton.onDidClick(() => {
 			this.sourceIsDacpac = false;
 			if ((this.sourceServerDropdown.value as ConnectionDropdownValue)) {
-				this.formBuilder.insertFormItem(this.sourceServerComponent, 1, { horizontal: true, componentWidth: 300 });
-				this.formBuilder.insertFormItem(this.sourceDatabaseComponent, 2, { horizontal: true, componentWidth: 300 });
+				this.formBuilder.insertFormItem(this.sourceServerComponent, 2, { horizontal: true, componentWidth: 300, titleFontSize: titleFontSize });
+				this.formBuilder.insertFormItem(this.sourceDatabaseComponent, 3, { horizontal: true, componentWidth: 300, titleFontSize: titleFontSize });
 			} else {
-				this.formBuilder.insertFormItem(this.sourceNoActiveConnectionsText, 1, { horizontal: true });
+				this.formBuilder.insertFormItem(this.sourceNoActiveConnectionsText, 2, { horizontal: true, titleFontSize: titleFontSize });
 			}
 			this.formBuilder.removeFormItem(this.sourceDacpacComponent);
 		});
@@ -285,7 +316,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: flexRadioButtonsModel,
-			title: SourceRadioButtonsLabel
+			title: RadioButtonsLabel
 		};
 	}
 
@@ -308,7 +339,7 @@ export class SchemaCompareDialog {
 			this.formBuilder.removeFormItem(this.targetNoActiveConnectionsText);
 			this.formBuilder.removeFormItem(this.targetServerComponent);
 			this.formBuilder.removeFormItem(this.targetDatabaseComponent);
-			this.formBuilder.addFormItem(this.targetDacpacComponent, { horizontal: true });
+			this.formBuilder.addFormItem(this.targetDacpacComponent, { horizontal: true, titleFontSize: titleFontSize });
 		});
 
 		// show server and db dropdowns or 'No active connections' text
@@ -316,10 +347,10 @@ export class SchemaCompareDialog {
 			this.targetIsDacpac = false;
 			this.formBuilder.removeFormItem(this.targetDacpacComponent);
 			if ((this.targetServerDropdown.value as ConnectionDropdownValue)) {
-				this.formBuilder.addFormItem(this.targetServerComponent, { horizontal: true, componentWidth: 300 });
-				this.formBuilder.addFormItem(this.targetDatabaseComponent, { horizontal: true, componentWidth: 300 });
+				this.formBuilder.addFormItem(this.targetServerComponent, { horizontal: true, componentWidth: 300, titleFontSize: titleFontSize });
+				this.formBuilder.addFormItem(this.targetDatabaseComponent, { horizontal: true, componentWidth: 300, titleFontSize: titleFontSize });
 			} else {
-				this.formBuilder.addFormItem(this.targetNoActiveConnectionsText, { horizontal: true });
+				this.formBuilder.addFormItem(this.targetNoActiveConnectionsText, { horizontal: true, titleFontSize: titleFontSize });
 			}
 		});
 
@@ -332,7 +363,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: flexRadioButtonsModel,
-			title: TargetRadioButtonsLabel
+			title: RadioButtonsLabel
 		};
 	}
 
@@ -344,7 +375,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: this.sourceServerDropdown,
-			title: localize('schemaCompare.sourceServerDropdownTitle', 'Source Server')
+			title: ServerDropdownLabel
 		};
 	}
 
@@ -356,7 +387,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: this.targetServerDropdown,
-			title: localize('schemaCompare.targetServerDropdownTitle', 'Target Server')
+			title: ServerDropdownLabel
 		};
 	}
 
@@ -376,26 +407,37 @@ export class SchemaCompareDialog {
 			return undefined;
 		}
 
+		let count = -1;
+		let idx = -1;
 		let values = cons.map(c => {
+			count++;
+
+			if (c.connectionId === this.connectionId) {
+				idx = count;
+			}
+
 			let db = c.options.databaseDisplayName;
 			let usr = c.options.user;
 			let srv = c.options.server;
 
-			if (!db) {
-				db = '<default>';
-			}
-
 			if (!usr) {
-				usr = 'default';
+				usr = localize('schemaCompareDialog.defaultUser', 'default');
 			}
 
-			let finalName = `${srv}, ${db} (${usr})`;
+			let finalName = `${srv} (${usr})`;
 			return {
 				connection: c,
 				displayName: finalName,
 				name: srv
 			};
 		});
+
+		// move server of current connection to the top of the list so it is the default
+		if (idx >= 1) {
+			let tmp = values[0];
+			values[0] = values[idx];
+			values[idx] = tmp;
+		}
 
 		return values;
 	}
@@ -405,7 +447,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: this.sourceDatabaseDropdown,
-			title: localize('schemaCompare.sourceDatabaseDropdownTitle', 'Source Database')
+			title: DatabaseDropdownLabel
 		};
 	}
 
@@ -414,7 +456,7 @@ export class SchemaCompareDialog {
 
 		return {
 			component: this.targetDatabaseDropdown,
-			title: localize('schemaCompare.targetDatabaseDropdownTitle', 'Target Database')
+			title: DatabaseDropdownLabel
 		};
 	}