+++ src/sql/azdata.proposed.d.ts
@@ -1731,8 +1731,8 @@ declare module 'azdata' {
 		Property = 1
 	}
 	export enum SchemaCompareEndpointType {
-		database = 0,
-		dacpac = 1
+		Database = 0,
+		Dacpac = 1
 	}
 	export interface SchemaCompareEndpointInfo {
 		endpointType: SchemaCompareEndpointType;
@@ -1742,10 +1742,166 @@ declare module 'azdata' {
 		ownerUri: string;
 	}
 
+	export interface SchemaCompareOptionsResult extends ResultStatus {
+		defaultDeploymentOptions: DeploymentOptions;
+	}
+
+	export interface DeploymentOptions {
+		ignoreTableOptions: boolean;
+		ignoreSemicolonBetweenStatements: boolean;
+		ignoreRouteLifetime: boolean;
+		ignoreRoleMembership: boolean;
+		ignoreQuotedIdentifiers: boolean;
+		ignorePermissions: boolean;
+		ignorePartitionSchemes: boolean;
+		ignoreObjectPlacementOnPartitionScheme: boolean;
+		ignoreNotForReplication: boolean;
+		ignoreLoginSids: boolean;
+		ignoreLockHintsOnIndexes: boolean;
+		ignoreKeywordCasing: boolean;
+		ignoreIndexPadding: boolean;
+		ignoreIndexOptions: boolean;
+		ignoreIncrement: boolean;
+		ignoreIdentitySeed: boolean;
+		ignoreUserSettingsObjects: boolean;
+		ignoreFullTextCatalogFilePath: boolean;
+		ignoreWhitespace: boolean;
+		ignoreWithNocheckOnForeignKeys: boolean;
+		verifyCollationCompatibility: boolean;
+		unmodifiableObjectWarnings: boolean;
+		treatVerificationErrorsAsWarnings: boolean;
+		scriptRefreshModule: boolean;
+		scriptNewConstraintValidation: boolean;
+		scriptFileSize: boolean;
+		scriptDeployStateChecks: boolean;
+		scriptDatabaseOptions: boolean;
+		scriptDatabaseCompatibility: boolean;
+		scriptDatabaseCollation: boolean;
+		runDeploymentPlanExecutors: boolean;
+		registerDataTierApplication: boolean;
+		populateFilesOnFileGroups: boolean;
+		noAlterStatementsToChangeClrTypes: boolean;
+		includeTransactionalScripts: boolean;
+		includeCompositeObjects: boolean;
+		allowUnsafeRowLevelSecurityDataMovement: boolean;
+		ignoreWithNocheckOnCheckConstraints: boolean;
+		ignoreFillFactor: boolean;
+		ignoreFileSize: boolean;
+		ignoreFilegroupPlacement: boolean;
+		doNotAlterReplicatedObjects: boolean;
+		doNotAlterChangeDataCaptureObjects: boolean;
+		disableAndReenableDdlTriggers: boolean;
+		deployDatabaseInSingleUserMode: boolean;
+		createNewDatabase: boolean;
+		compareUsingTargetCollation: boolean;
+		commentOutSetVarDeclarations: boolean;
+		blockWhenDriftDetected: boolean;
+		blockOnPossibleDataLoss: boolean;
+		backupDatabaseBeforeChanges: boolean;
+		allowIncompatiblePlatform: boolean;
+		allowDropBlockingAssemblies: boolean;
+		dropConstraintsNotInSource: boolean;
+		dropDmlTriggersNotInSource: boolean;
+		dropExtendedPropertiesNotInSource: boolean;
+		dropIndexesNotInSource: boolean;
+		ignoreFileAndLogFilePath: boolean;
+		ignoreExtendedProperties: boolean;
+		ignoreDmlTriggerState: boolean;
+		ignoreDmlTriggerOrder: boolean;
+		ignoreDefaultSchema: boolean;
+		ignoreDdlTriggerState: boolean;
+		ignoreDdlTriggerOrder: boolean;
+		ignoreCryptographicProviderFilePath: boolean;
+		verifyDeployment: boolean;
+		ignoreComments: boolean;
+		ignoreColumnCollation: boolean;
+		ignoreAuthorizer: boolean;
+		ignoreAnsiNulls: boolean;
+		generateSmartDefaults: boolean;
+		dropStatisticsNotInSource: boolean;
+		dropRoleMembersNotInSource: boolean;
+		dropPermissionsNotInSource: boolean;
+		dropObjectsNotInSource: boolean;
+		ignoreColumnOrder: boolean;
+		doNotDropObjectTypes: SchemaObjectType[];
+		excludeObjectTypes: SchemaObjectType[];
+	}
+
+	export enum SchemaObjectType {
+		Aggregates = 0,
+		ApplicationRoles = 1,
+		Assemblies = 2,
+		AssemblyFiles = 3,
+		AsymmetricKeys = 4,
+		BrokerPriorities = 5,
+		Certificates = 6,
+		ColumnEncryptionKeys = 7,
+		ColumnMasterKeys = 8,
+		Contracts = 9,
+		DatabaseOptions = 10,
+		DatabaseRoles = 11,
+		DatabaseTriggers = 12,
+		Defaults = 13,
+		ExtendedProperties = 14,
+		ExternalDataSources = 15,
+		ExternalFileFormats = 16,
+		ExternalTables = 17,
+		Filegroups = 18,
+		FileTables = 19,
+		FullTextCatalogs = 20,
+		FullTextStoplists = 21,
+		MessageTypes = 22,
+		PartitionFunctions = 23,
+		PartitionSchemes = 24,
+		Permissions = 25,
+		Queues = 26,
+		RemoteServiceBindings = 27,
+		RoleMembership = 28,
+		Rules = 29,
+		ScalarValuedFunctions = 30,
+		SearchPropertyLists = 31,
+		SecurityPolicies = 32,
+		Sequences = 33,
+		Services = 34,
+		Signatures = 35,
+		StoredProcedures = 36,
+		SymmetricKeys = 37,
+		Synonyms = 38,
+		Tables = 39,
+		TableValuedFunctions = 40,
+		UserDefinedDataTypes = 41,
+		UserDefinedTableTypes = 42,
+		ClrUserDefinedTypes = 43,
+		Users = 44,
+		Views = 45,
+		XmlSchemaCollections = 46,
+		Audits = 47,
+		Credentials = 48,
+		CryptographicProviders = 49,
+		DatabaseAuditSpecifications = 50,
+		DatabaseEncryptionKeys = 51,
+		DatabaseScopedCredentials = 52,
+		Endpoints = 53,
+		ErrorMessages = 54,
+		EventNotifications = 55,
+		EventSessions = 56,
+		LinkedServerLogins = 57,
+		LinkedServers = 58,
+		Logins = 59,
+		MasterKeys = 60,
+		Routes = 61,
+		ServerAuditSpecifications = 62,
+		ServerRoleMembership = 63,
+		ServerRoles = 64,
+		ServerTriggers = 65
+	}
+
 	export interface SchemaCompareServicesProvider extends DataProvider {
-		schemaCompare(sourceEndpointInfo: SchemaCompareEndpointInfo, targetEndpointInfo: SchemaCompareEndpointInfo, taskExecutionMode: TaskExecutionMode): Thenable<SchemaCompareResult>;
+		schemaCompare(sourceEndpointInfo: SchemaCompareEndpointInfo, targetEndpointInfo: SchemaCompareEndpointInfo, taskExecutionMode: TaskExecutionMode, deploymentOptions: DeploymentOptions): Thenable<SchemaCompareResult>;
 		schemaCompareGenerateScript(operationId: string, targetDatabaseName: string, scriptFilePath: string, taskExecutionMode: TaskExecutionMode): Thenable<ResultStatus>;
 		schemaComparePublishChanges(operationId: string, targetServerName: string, targetDatabaseName: string, taskExecutionMode: TaskExecutionMode): Thenable<ResultStatus>;
+		schemaCompareGetDefaultOptions(): Thenable<SchemaCompareOptionsResult>;
+		schemaCompareIncludeExcludeNode(operationId: string, diffEntry: DiffEntry, IncludeRequest: boolean, taskExecutionMode: TaskExecutionMode): Thenable<ResultStatus>;
 	}
 
 	// Security service interfaces ------------------------------------------------------------------------
@@ -2171,11 +2327,11 @@ declare module 'azdata' {
 	 * AccountProvider.refresh or AccountProvider.prompt are rejected with this error, the error
 	 * will not be reported to the user.
 	 */
-	export interface UserCancelledSignInError extends Error {
+	export interface PromptFailedResult {
 		/**
 		 * Type guard for differentiating user cancelled sign in errors from other errors
 		 */
-		userCancelledSignIn: boolean;
+		canceled: boolean;
 	}
 
 	/**
@@ -2226,7 +2382,7 @@ declare module 'azdata' {
 		 * Prompts the user to enter account information.
 		 * Returns an error if the user canceled the operation.
 		 */
-		prompt(): Thenable<Account>;
+		prompt(): Thenable<Account | PromptFailedResult>;
 
 		/**
 		 * Refreshes a stale account.
@@ -2234,7 +2390,7 @@ declare module 'azdata' {
 		 * Otherwise, returns a new updated account instance.
 		 * @param account - An account.
 		 */
-		refresh(account: Account): Thenable<Account>;
+		refresh(account: Account): Thenable<Account | PromptFailedResult>;
 
 		/**
 		 * Clears sensitive information for an account. To be called when account is removed
@@ -3966,7 +4122,7 @@ declare module 'azdata' {
 		 * Opens the connection and add it to object explorer and opens the dashboard and returns the ConnectionResult
 		 * @param connectionProfile connection profile
 		 */
-		export function connect(connectionProfile: IConnectionProfile): Thenable<ConnectionResult>;
+		export function connect(connectionProfile: IConnectionProfile, saveConnection?: boolean, showDashboard?: boolean): Thenable<ConnectionResult>;
 	}
 
 	export namespace nb {
@@ -4960,6 +5116,10 @@ declare module 'azdata' {
 		 */
 		export interface IStdinMessage extends IMessage {
 			channel: 'stdin';
+			content: {
+				prompt: string;
+				password: boolean;
+			};
 		}
 
 		/**