+++ src/sql/platform/dialog/dialogPane.ts
@@ -3,8 +3,6 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-'use strict';
-
 import 'vs/css!./media/dialogModal';
 
 import { NgModuleRef } from '@angular/core';
@@ -21,6 +19,8 @@ import { IThemable } from 'vs/platform/theme/common/styler';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { Emitter } from 'vs/base/common/event';
+import { attachTabbedPanelStyler } from 'sql/platform/theme/common/styler';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
 
 export class DialogPane extends Disposable implements IThemable {
 	private _tabbedPanel: TabbedPanel;
@@ -40,6 +40,7 @@ export class DialogPane extends Disposable implements IThemable {
 		private _content: string | DialogTab[],
 		private _validityChangedCallback: (valid: boolean) => void,
 		private _instantiationService: IInstantiationService,
+		private _themeService: IThemeService,
 		public displayPageTitle: boolean,
 		public description?: string,
 	) {
@@ -53,6 +54,7 @@ export class DialogPane extends Disposable implements IThemable {
 			this.initializeModelViewContainer(this._body, modelViewId);
 		} else {
 			this._tabbedPanel = new TabbedPanel(this._body);
+			attachTabbedPanelStyler(this._tabbedPanel, this._themeService);
 			this._content.forEach((tab, tabIndex) => {
 				if (this._selectedTabIndex === tabIndex) {
 					this._selectedTabContent = tab.content;