+++ src/sql/workbench/parts/dashboard/widgets/insights/views/charts/types/timeSeriesChart.contribution.ts
@@ -3,8 +3,7 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { mixin } from 'vs/base/common/objects';
-import { clone } from 'sql/base/common/objects';
+import { mixin, deepClone } from 'vs/base/common/objects';
 import { IJSONSchema } from 'vs/base/common/jsonSchema';
 
 import { registerInsight } from 'sql/platform/dashboard/common/insightRegistry';
@@ -15,6 +14,6 @@ import TimeSeriesChart from './timeSeriesChart.component';
 const properties: IJSONSchema = {
 };
 
-const timeSeriesSchema = mixin(clone(barChartSchema), properties) as IJSONSchema;
+const timeSeriesSchema = mixin(deepClone(barChartSchema), properties) as IJSONSchema;
 
 registerInsight('timeSeries', '', timeSeriesSchema, TimeSeriesChart);