+++ src/sqltest/parts/commandLine/commandLineService.test.ts
@@ -13,11 +13,8 @@ import * as Constants from 'sql/platform/connection/common/constants';
 import { ParsedArgs } from 'vs/platform/environment/common/environment';
 import { ICapabilitiesService } from 'sql/platform/capabilities/common/capabilitiesService';
 import { CapabilitiesTestService } from 'sqltest/stubs/capabilitiesTestService';
-import { QueryEditorService } from 'sql/workbench/services/queryEditor/browser/queryEditorService';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
-import { ObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
 import { IConnectionManagementService, IConnectionCompletionOptions, ConnectionType } from 'sql/platform/connection/common/connectionManagement';
-import { ConnectionStore } from 'sql/platform/connection/common/connectionStore';
 import { TestConnectionManagementService } from 'sqltest/stubs/connectionManagementService.test';
 import { ICommandService } from 'vs/platform/commands/common/commands';
 import { TestCommandService } from 'vs/editor/test/browser/editorTestServices';
@@ -25,9 +22,10 @@ import { WorkspaceConfigurationTestService } from 'sqltest/stubs/workspaceConfig
 import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';
 import { assertThrowsAsync } from 'sqltest/utils/testUtils';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { TestEditorService } from 'vs/workbench/test/workbenchTestServices';
+import { TestEditorService, TestLogService } from 'vs/workbench/test/workbenchTestServices';
 import { QueryInput } from 'sql/workbench/parts/query/common/queryInput';
 import { URI } from 'vs/base/common/uri';
+import { ILogService } from 'vs/platform/log/common/log';
 
 class TestParsedArgs implements ParsedArgs {
 	[arg: string]: any;
@@ -95,22 +93,17 @@ class TestParsedArgs implements ParsedArgs {
 suite('commandLineService tests', () => {
 
 	let capabilitiesService: CapabilitiesTestService;
-	let commandLineService: CommandLineService;
-	let queryEditorService: TypeMoq.Mock<QueryEditorService>;
-	let editorService: TypeMoq.Mock<IEditorService>;
-	let objectExplorerService: TypeMoq.Mock<ObjectExplorerService>;
-	let connectionStore: TypeMoq.Mock<ConnectionStore>;
 
 	setup(() => {
 		capabilitiesService = new CapabilitiesTestService();
-		connectionStore = TypeMoq.Mock.ofType(ConnectionStore);
 	});
 
 	function getCommandLineService(connectionManagementService: IConnectionManagementService,
 		configurationService: IConfigurationService,
 		capabilitiesService?: ICapabilitiesService,
 		commandService?: ICommandService,
-		editorService?: IEditorService
+		editorService?: IEditorService,
+		logService?: ILogService
 	): CommandLineService {
 		let service = new CommandLineService(
 			capabilitiesService,
@@ -121,7 +114,8 @@ suite('commandLineService tests', () => {
 			editorService,
 			commandService,
 			configurationService,
-			undefined
+			undefined,
+			logService
 		);
 		return service;
 	}
@@ -205,7 +199,8 @@ suite('commandLineService tests', () => {
 			.verifiable(TypeMoq.Times.once());
 		connectionManagementService.setup(c => c.getConnectionProfileById(TypeMoq.It.isAnyString())).returns(() => originalProfile);
 		const configurationService = getConfigurationServiceMock(true);
-		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService);
+		const logService = new TestLogService();
+		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService, undefined, undefined, logService);
 		await service.processCommandLine(args);
 		connectionManagementService.verifyAll();
 	});
@@ -309,7 +304,8 @@ suite('commandLineService tests', () => {
 		connectionManagementService.setup(c => c.getConnectionProfileById(TypeMoq.It.isAnyString())).returns(() => originalProfile);
 		connectionManagementService.setup(c => c.getConnectionGroups(TypeMoq.It.isAny())).returns(() => []);
 		const configurationService = getConfigurationServiceMock(true);
-		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService);
+		const logService = new TestLogService();
+		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService, undefined, undefined, logService);
 		await service.processCommandLine(args);
 		connectionManagementService.verifyAll();
 	});
@@ -351,7 +347,8 @@ suite('commandLineService tests', () => {
 		connectionManagementService.setup(c => c.getConnectionProfileById('testID')).returns(() => originalProfile).verifiable(TypeMoq.Times.once());
 		connectionManagementService.setup(x => x.getConnectionGroups(TypeMoq.It.isAny())).returns(() => [conProfGroup]);
 		const configurationService = getConfigurationServiceMock(true);
-		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService);
+		const logService = new TestLogService();
+		let service = getCommandLineService(connectionManagementService.object, configurationService.object, capabilitiesService, undefined, undefined, logService);
 		await service.processCommandLine(args);
 		connectionManagementService.verifyAll();
 	});