+++ src/sqltest/parts/jobManagement/jobManagementService.test.ts
@@ -3,7 +3,6 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-'use strict';
 import { JobManagementService } from 'sql/platform/jobManagement/common/jobManagementService';
 
 // TESTS ///////////////////////////////////////////////////////////////////